Refactor folder structure of CXX and FII bindings
