+++ bindings/wasm/native/cxx_bridge/gguf_callbacks.h
