+++ crates/sys/native/cxx_bridge/cogent_cxx.cpp
