+++ packages/npm/src/wasm/wasm-bridge.ts
@@ -31,8 +31,8 @@ import { withDerivedObservabilityMetrics } from '../engine/inference-types.js';
 import { createAbortError } from '../utils/abort.js';
 import { assertGrammarByteSize } from '../utils/grammar.js';
 
-// Mirror of CE_TokenEmissionMode in native/api/ffi_types.h.  Native exposes
-// only NONE (no emission) and STREAMING_BUFFER (SAB ring).
+// Mirror of CE_TokenEmissionMode in bindings/wasm/native/js_api/ffi_types.h.
+// Native exposes only NONE (no emission) and STREAMING_BUFFER (SAB ring).
 export const TOKEN_EMISSION_NONE = 0;
 export const TOKEN_EMISSION_STREAMING_BUFFER = 1;
